[fix][client] Adjust code struct and fixup known issues.
