+++ src/client/common/dynamic_config.cpp
@@ -90,6 +90,79 @@ DEFINE_int32(disk_check_duration_millsecond, 1 * 1000,
              "disk health check duration in millsecond");
 DEFINE_validator(disk_check_duration_millsecond, &PassInt32);
 
+// thread num or bthread num
+DEFINE_uint32(stat_timer_thread_num, 8, "stat timer thread num");
+DEFINE_validator(stat_timer_thread_num, &PassUint32);
+
+DEFINE_uint32(fs_usage_flush_interval_second, 20,
+              "fs usage flush interval in second");
+DEFINE_validator(fs_usage_flush_interval_second, &PassUint32);
+
+DEFINE_uint32(flush_quota_interval_second, 30,
+              "flush quota interval in second");
+DEFINE_validator(flush_quota_interval_second, &PassUint32);
+
+DEFINE_uint32(load_quota_interval_second, 30, "flush quota interval in second");
+DEFINE_validator(load_quota_interval_second, &PassUint32);
+
+DEFINE_uint32(push_metric_interval_millsecond, 500,
+              "push client metrics interval in millsecond");
+DEFINE_validator(push_metric_interval_millsecond, &PassUint32);
+
+// fuse
+// kernal will retry when read fail
+DEFINE_uint32(fuse_read_max_retry_s3_not_exist, 60,
+              "fuse read max retry when s3 object not exist");
+DEFINE_validator(fuse_read_max_retry_s3_not_exist, &PassUint32);
+
+// from config.h and config.cpp
+DEFINE_bool(enableCto, true, "acheieve cto consistency");
+DEFINE_bool(useFakeS3, false,
+            "Use fake s3 to inject more metadata for testing metaserver");
+DEFINE_bool(supportKVcache, false, "use kvcache to speed up sharing");
+
+/**
+ * use curl -L fuseclient:port/flags/fuseClientAvgWriteBytes?setvalue=true
+ * for dynamic parameter configuration
+ */
+DEFINE_uint64(fuseClientAvgWriteBytes, 0,
+              "the write throttle bps of fuse client");
+DEFINE_validator(fuseClientAvgWriteBytes, &PassUint64);
+DEFINE_uint64(fuseClientBurstWriteBytes, 0,
+              "the write burst bps of fuse client");
+DEFINE_validator(fuseClientBurstWriteBytes, &PassUint64);
+DEFINE_uint64(fuseClientBurstWriteBytesSecs, 180,
+              "the times that write burst bps can continue");
+DEFINE_validator(fuseClientBurstWriteBytesSecs, &PassUint64);
+
+DEFINE_uint64(fuseClientAvgWriteIops, 0,
+              "the write throttle iops of fuse client");
+DEFINE_validator(fuseClientAvgWriteIops, &PassUint64);
+DEFINE_uint64(fuseClientBurstWriteIops, 0,
+              "the write burst iops of fuse client");
+DEFINE_validator(fuseClientBurstWriteIops, &PassUint64);
+DEFINE_uint64(fuseClientBurstWriteIopsSecs, 180,
+              "the times that write burst iops can continue");
+DEFINE_validator(fuseClientBurstWriteIopsSecs, &PassUint64);
+
+DEFINE_uint64(fuseClientAvgReadBytes, 0,
+              "the Read throttle bps of fuse client");
+DEFINE_validator(fuseClientAvgReadBytes, &PassUint64);
+DEFINE_uint64(fuseClientBurstReadBytes, 0, "the Read burst bps of fuse client");
+DEFINE_validator(fuseClientBurstReadBytes, &PassUint64);
+DEFINE_uint64(fuseClientBurstReadBytesSecs, 180,
+              "the times that Read burst bps can continue");
+DEFINE_validator(fuseClientBurstReadBytesSecs, &PassUint64);
+
+DEFINE_uint64(fuseClientAvgReadIops, 0,
+              "the Read throttle iops of fuse client");
+DEFINE_validator(fuseClientAvgReadIops, &PassUint64);
+DEFINE_uint64(fuseClientBurstReadIops, 0, "the Read burst iops of fuse client");
+DEFINE_validator(fuseClientBurstReadIops, &PassUint64);
+DEFINE_uint64(fuseClientBurstReadIopsSecs, 180,
+              "the times that Read burst iops can continue");
+DEFINE_validator(fuseClientBurstReadIopsSecs, &PassUint64);
+
 }  // namespace common
 }  // namespace client
 }  // namespace dingofs
