+++ src/client/datastream/CMakeLists.txt
@@ -19,6 +19,5 @@ add_library(client_datastream
     page_allocator.cpp
 )
 target_link_libraries(client_datastream
-    vfs_old_common
     dingofs_utils
 )
