+++ src/client/vfs/vfs_meta.h
@@ -69,9 +69,6 @@ struct Attr {
   uint64_t atime{0};
   uint64_t mtime{0};
   uint64_t ctime{0};
-  // uint32_t mtime_ns{0};
-  // uint32_t atime_ns{0};
-  // uint32_t ctime_ns{0};
   FileType type;
   // TODO: refact, maybe use separate key for hardlink
   std::vector<Ino> parents;
