--- src/client/vfs_old/common/CMakeLists.txt
@@ -1,32 +1,0 @@
-# Copyright (c) 2024 dingodb.com, Inc. All Rights Reserved
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# Define the BASE_FLAGS and DINGO_DEFAULT_COPTS variables
-
-add_library(vfs_old_common
-    common.cpp
-    config.cpp
-    dynamic_config.cpp
-)
-
-target_link_libraries(vfs_old_common
-    PROTO_OBJS
-    stub_common
-    aws_s3_adapter
-    dingofs_utils
-    dingofs_base_lib
-    client_common
-    brpc::brpc
-    gflags::gflags
-    glog::glog
-)
