+++ src/client/vfs_old/filesystem/dir_quota_manager.cpp
@@ -18,11 +18,11 @@
 #include <cstdint>
 #include <unordered_map>
 
-#include "client/vfs_old/common/dynamic_config.h"
+#include "client/common/dynamic_config.h"
 #include "client/vfs_old/inode_wrapper.h"
 #include "common/define.h"
-#include "utils/concurrent/concurrent.h"
 #include "glog/logging.h"
+#include "utils/concurrent/concurrent.h"
 
 namespace dingofs {
 namespace client {
