+++ src/client/vfs_old/filesystem/lookup_cache.h
@@ -26,7 +26,7 @@
 #include <memory>
 #include <string>
 
-#include "client/vfs_old/common/config.h"
+#include "client/common/config.h"
 #include "client/vfs_old/filesystem/meta.h"
 #include "utils/lru_cache.h"
