+++ src/client/vfs_old/kvclient/CMakeLists.txt
@@ -20,7 +20,6 @@ add_library(client_memcached_client
 
 target_link_libraries(client_memcached_client
     PROTO_OBJS
-    vfs_old_common
     stub_metric
     dingofs_utils
     aws_s3_adapter
