+++ src/client/vfs_old/lease/CMakeLists.txt
@@ -19,7 +19,6 @@ add_library(client_lease
 target_link_libraries(client_lease
     PROTO_OBJS
     rpcclient
-    vfs_old_common
     brpc::brpc
     gflags::gflags
     glog::glog
